Created file; Added massmsg command; Altered invite post prevention;
Created massmsg.js; Added massmsg command to massmsg.js, which messages the specified message to all members of the guild who it can reach so long as the command executor and the bot both have the "ADMINISTRATOR" permission, and logs said message, its sender's username, and the amount of users it attempted to message; Altered invite post prevention by allowing anyone with the "ADMINISTRATOR" permission to post invites by returning the event handling function;

+++ bot.js
@@ -31,7 +31,9 @@ client.on("message", (message) => {
   var content = message.content;
 
   if(/discord\.gg\//.test(content) || /\.gg\/[a-zA-Z0-9]/.test(content)) {
-// || /\.gg/.test(content)
+    if(message.guild.member(message.author).hasPermission("ADMINISTRATOR")) {
+      return;
+    }
     message.delete();
     message.reply("please refrain from posting invite links.");
     let channel = message.guild.channels.find('name', 'mod_logs');
